+++ src/com/avona/games/towerdefence/StationaryObject.java
@@ -17,7 +17,7 @@ public abstract class StationaryObject {
 	 */
 	public abstract void step(final float dt);
 
-	public boolean collidesWith(final V2 otherLocation, final double otherRadius) {
+	public boolean collidesWith(final V2 otherLocation, final float otherRadius) {
 		return Collision.circleCollidesWithCircle(location, radius,
 				otherLocation, otherRadius);
 	}
